backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Abhijit Gupta <avgupta456@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
openai = "^0.2.4"
flask = "^1.1.2"
flask_cors = "^3.0.9"
python-dotenv = "^0.14.0"
google-cloud-speech = "^1.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
